cleanned component with shared lib in common folder
