+++ services/ui-src/src/views/Qualifiers/Adult/index.tsx
@@ -94,21 +94,7 @@ export const ACSQualifiers = () => {
   };
 
   const validateAndSetErrors = (data: ACSQualifierForm): boolean => {
-    const validationErrors = validationFunctions.reduce(
-      (acc: any, current: any) => {
-        const error = current(data);
-        let errorArray = [];
-
-        if (Array.isArray(error)) {
-          errorArray = [...error];
-        } else {
-          errorArray = [error];
-        }
-
-        return error ? [...acc, ...errorArray] : acc;
-      },
-      []
-    );
+    const validationErrors = Common.validateData(validationFunctions, data);
     setErrors(validationErrors.length > 0 ? validationErrors : undefined);
     return validationErrors.length > 0;
   };
